additions to the test runner to accomodate global loggers
the library code assumes an externed 'MainLogger' symbol that causes the
tests to fail to link. as an alternative to adding DI everywhere for
logging, i opted to define that symbol in the tests as well. by default,
it is configured off, but can be enabled with a --verbose flag for
test debugging purposes in order to force messages to stdout.