restructuring the driveshaft build
now builds a lib primarily, which the driveshaft binary links against.
the test executable links against the same library.

of note is the switch from glob() to set(). i'm new to CMake, but a
brief reading of the available literature suggests that the latter
is preferable since glob()'ing hides dependency changes from the build system.