first pass at adding metrics library
this adds a simple metrics library that makes it possible to track counts and
gauges for now. It makes it accessible as a variable directly in the
Driveshaft namespace. And the status loop has been extended to print all
counters and gauges for the "counters" and "gauges" commands respectively.

As an example it only tracks pool size for now, which is kinda redundant to
the "threads" command but a good example.