Let compiler pick C++ standard instead of limiting to 11
The newer version of log4cxx we're pulling in assumes
std::shared_mutex is available which needs C++17. Not sure how to set
a minimum std version.

This was causing build errors on Ubuntu 22.04 as shared_ptr is
available in C++17 not 11. Not sure what how this was working before.